fix_tensor_mismatch.py
#!/usr/bin/env python3
"""
Fix for tensor size mismatch error in ring attractor network test files.

This script documents and fixes the tensor size mismatch issue where:
- Model expects external input of size (n_exc,) = (800,) 
- But test files were creating dummy inputs of size (64,)

The error was:
RuntimeError: The size of tensor a (800) must match the size of tensor b (64) at non-singleton dimension 0

Fixed by updating test files to use correct input dimensions.
"""

import torch
from hd_ring_attractor.src.models import RingAttractorNetwork


def demonstrate_tensor_size_fix():
    """
    Demonstrate the correct way to create inputs for the ring attractor network.
    """
    print("Ring Attractor Network - Tensor Size Fix Demonstration")
    print("=" * 60)
    
    # Create model with standard parameters
    device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
    model = RingAttractorNetwork(n_exc=800, n_inh=200, device=device)
    
    print(f"Model configuration:")
    print(f"  - Excitatory neurons: {model.n_exc}")
    print(f"  - Inhibitory neurons: {model.n_inh}")
    print(f"  - Device: {device}")
    
    # CORRECT: External input must match number of excitatory neurons
    print(f"\n✓ CORRECT input size:")
    correct_input = torch.zeros(model.n_exc, device=device)
    print(f"  Input shape: {correct_input.shape}")
    print(f"  Expected shape: ({model.n_exc},)")
    
    try:
        output = model(correct_input, steps=1)
        print(f"  Forward pass successful!")
        print(f"  Output shape: {output.shape}")
    except Exception as e:
        print(f"  Error: {e}")
    
    # INCORRECT: Shows what was causing the error
    print(f"\n✗ INCORRECT input size (fixed):")
    incorrect_input = torch.zeros(64, device=device)
    print(f"  Input shape: {incorrect_input.shape}")
    print(f"  Expected shape: ({model.n_exc},)")
    print(f"  This would cause: RuntimeError - tensor size mismatch")
    
    print(f"\n📋 Summary of fixes applied:")
    print(f"  - test_poisson_noise.py: Updated dummy_input from size 64 to 800")
    print(f"  - test_poisson_noise.py: Updated target tensor from size 64 to 800")
    print(f"  - All external inputs must match model.n_exc (800) neurons")


if __name__ == "__main__":
    demonstrate_tensor_size_fix()
